track/serato/markers2: Move Serato Markers2 code into serato directory
