+++ src/test/serato/markers2-data/hotcue-00m00s0-blue.octet-stream
