+++ src/test/serato/markers2-data/hotcue-colors.octet-stream
