+++ src/test/serato/markers2-data/hotcues-with-names.octet-stream
